pyproject.toml
[tool.poetry]
name = "anomaly-detector"
version = "0.1.0"
description = ""
authors = ["tnahddisttud <siddhantp73@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"
pandas = "^2.2.2"
numpy = "^1.26.4"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"
statsmodels = "^0.14.2"
scikit-learn = "^1.4.2"
pyspark = "^3.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
